RAW copy of tivobuddy, a program to pull shows off of one or more tivos in your home, re-encode them as mp4, and copy to a directory of your choosing.  It finds the tivos using avahi, so you dont have to worry about their ip addresses changing.  You just refer to them by name.  It also has a concept of storing the show lists in a cache, so that you can later schedule the download of the show without having to rescan.
This is one of my earliest prototypes.  In it's current checked in form it will not work, as I have removed references to my tivo Media Access Key.  I also have not included details on the mysql database this package can use or how to use each of the attached utilities.

I plan to update this repo with better documentation :)